+++ stoix/networks/postprocessors.py
@@ -66,7 +66,9 @@ class ScalePostProcessor(nn.Module):
 
     @nn.compact
     def __call__(self, distribution: Distribution) -> Distribution:
-        post_processor = partial(self.scale_fn, minimum=self.minimum, maximum=self.maximum)
+        post_processor = partial(
+            self.scale_fn, minimum=self.minimum, maximum=self.maximum
+        )  # type: ignore
         return PostProcessedDistribution(distribution, post_processor)
