+++ stoix/systems/ppo/anakin/ff_ppo_penalty_continuous.py
@@ -577,7 +577,7 @@ def run_experiment(_config: DictConfig) -> float:
 
 
 @hydra.main(
-    config_path="../../configs", config_name="default_ff_ppo_continuous.yaml", version_base="1.2"
+    config_path="../../../configs", config_name="default_ff_ppo_continuous.yaml", version_base="1.2"
 )
 def hydra_entry_point(cfg: DictConfig) -> float:
     """Experiment entry point."""
